completely move nfc package to libnfc module
